Rename the "vmware" directory to "vsphere".
