+++ provider/all/all.go
@@ -14,5 +14,5 @@ import (
 	_ "github.com/juju/juju/provider/maas"
 	_ "github.com/juju/juju/provider/manual"
 	_ "github.com/juju/juju/provider/openstack"
-	_ "github.com/juju/juju/provider/vmware"
+	_ "github.com/juju/juju/provider/vsphere"
 )
