+++ provider/vsphere/client.go
