+++ provider/vsphere/config.go
