+++ provider/vsphere/environ.go
