+++ provider/vsphere/environ_availzones.go
