+++ provider/vsphere/environ_policy.go
