+++ provider/vsphere/fake_methods_test.go
