+++ provider/vsphere/image_metadata.go
