+++ provider/vsphere/ovf_import_manager.go
