+++ provider/vsphere/package_test.go
