+++ provider/vsphere/testing_test.go
